Invoke-WebRequest -Resume
I decided not to plug `IWR` into this because `IWR` is not meant for downloading files and this parameter set is officially considered a mistake of the evolution and a dead end.  However, once we have arrived here, be so kind and add a `-Resume` for the poor guys on the other end of the slow commuted link, will you?  BTW, sorry for not having tested what the download contains.